services/academy/backend/requirements.txt
# =======================
# Core FastAPI and backend
# =======================
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-multipart==0.0.6
pydantic==2.11.7
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
fastapi-users[sqlalchemy2]==12.1.2
bcrypt==4.1.2
PyMuPDF==1.24.2

# =======================
# PDF parsing & Transformers
# =======================
PyPDF2==3.0.1
transformers==4.37.2
tokenizers==0.15.2
sentence-transformers==2.6.1

# =======================
# LlamaIndex (0.12.x stack)
# =======================
llama-index==0.12.45
llama-index-core==0.12.45
llama-index-llms-ollama==0.4.0
ollama>=0.3.0,<0.5.0
llama-index-embeddings-huggingface==0.4.0
llama-index-vector-stores-chroma==0.4.2
llama-index-agent-openai==0.4.12
llama-index-llms-openai==0.4.7
llama-index-question-gen-openai==0.3.1
llama-index-readers-file==0.4.0
# llama-index-multi-modal-llms-openai==0.5.0
llama-index-program-openai==0.3.2
llama-index-cli==0.4.2
# llama-index-indices-managed-llama-cloud==0.4.2
llama-index-embeddings-openai==0.3.1


# =======================
# Vector DB (ChromaDB)
# =======================
chromadb==1.0.13

# =======================
# Ollama & OpenAI SDK
# =======================
# keep Ollama pinned below 0.1.7 to avoid httpx conflict
openai==1.97.0

# =======================
# Utilities
# =======================
tqdm==4.67.1
requests==2.32.3
filelock==3.18.0
fsspec==2025.5.1
packaging==25.0
pyyaml==6.0.2
typing-extensions==4.14.0
colorama==0.4.6
scikit-learn==1.4.2
numpy==1.26.4
edge-tts
